requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
neo4j>=5.14.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
sentence-transformers>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.1.0
tqdm>=4.66.0
